+++ lib/active_record/pg_enum/5.1/prepare_column_options.rb
@@ -1,0 +1,1 @@
+require "active_record/pg_enum/5.0/prepare_column_options"