+++ lib/active_record/pg_enum/5.1/schema_dumper.rb
@@ -1,0 +1,1 @@
+require "active_record/pg_enum/5.0/schema_dumper"