+++ lib/active_record/pg_enum/5.2/schema_dumper.rb
@@ -1,0 +1,10 @@
+require "active_record/pg_enum/5.0/schema_dumper"
+
+module ActiveRecord
+  module PGEnum
+    def self.install_schema_dumper
+      require "active_record/connection_adapters/postgresql/schema_dumper"
+      ActiveRecord::ConnectionAdapters::PostgreSQL::SchemaDumper.prepend SchemaDumper
+    end
+  end
+end